major file changes and updates
