--- macos_disks.py
@@ -1,7 +1,0 @@
-import subprocess
-
-def get_disks_darwin():
-    cmd = ["diskutil", "list"]
-    result = subprocess.run(cmd, capture_output=True, text=True)
-    return result.stdout   
-
